+++ apiserver/api/schemas/__init__.py
@@ -2,9 +2,11 @@
 # -*- coding: utf-8 -*-
 """Initializes schemas for all database collections required by Flask
 resources."""
-# First-party
-from apiserver.api.schemas.roles import RoleSchema
+# pylint: disable=R0401
 
-__all__ = [
-    'RoleSchema',
-]
+# First-party
+
+from apiserver.api.schemas.roles import RoleSchema
+from apiserver.api.schemas.users import UserSchema
+
+__all__ = ['RoleSchema', 'UserSchema']